Abstracted egg storage out of Data.java
Created EggStorageState to serve as an immutable record of how the egg is stored.
This simplifies egg storage and data validation, and will make it possible to easily convey how the egg is stored when calling events.
This also greatly simplified handling of egg storage types by classes that depend on Data.java.